perf: gradient checkpointing, bf16 pe_embedding, cross-attn refactor
- Add gradient_checkpointing config option for FlashTransformerEncoder
  and LatentTokenizer cross-attention (saves ~18GB at cost of 16% throughput)
- Store pe_embedding in bf16 to eliminate fp32→bf16 dtype conversions
  (2% speedup, 5.6GB memory saved)
- Cast ESM2 proj cache to bf16 for consistent dtype
- Extract _cross_attention() method for checkpoint compatibility
- Wire gradient_checkpointing through both SentenceTokenizer and LatentTokenizer

+++ src/state/emb/train/trainer.py
@@ -141,8 +141,9 @@ def main(cfg):
     model.collater = val_collator
     model = model.cuda()
 
-    # Load frozen protein embeddings onto the tokenizer
+    # Load frozen protein embeddings onto the tokenizer (bf16 to avoid dtype conversion)
     all_pe = get_embeddings(cfg)
+    all_pe = all_pe.to(torch.bfloat16)
     all_pe.requires_grad = False
     model.tokenizer.pe_embedding = nn.Embedding.from_pretrained(all_pe)
